extract interface for services that support the shard contracts
