+++ manager-parent/manager-v2/src/main/java/com/redhat/service/smartevents/manager/v2/api/user/SourceConnectorsAPI.java
@@ -38,7 +38,7 @@ import com.redhat.service.smartevents.manager.v2.api.user.models.requests.Connec
 import com.redhat.service.smartevents.manager.v2.api.user.models.responses.SourceConnectorListResponse;
 import com.redhat.service.smartevents.manager.v2.api.user.models.responses.SourceConnectorResponse;
 import com.redhat.service.smartevents.manager.v2.persistence.models.Connector;
-import com.redhat.service.smartevents.manager.v2.services.SourceConnectorService;
+import com.redhat.service.smartevents.manager.v2.services.SourceConnectorServiceImpl;
 
 import io.quarkus.security.Authenticated;
 
@@ -57,7 +57,7 @@ import io.quarkus.security.Authenticated;
 public class SourceConnectorsAPI {
 
     @Inject
-    SourceConnectorService sourceConnectorService;
+    SourceConnectorServiceImpl sourceConnectorService;
 
     @V2
     @Inject
