+++ manager-parent/manager-v2/src/main/java/com/redhat/service/smartevents/manager/v2/services/BridgeService.java
@@ -1,16 +1,13 @@
 package com.redhat.service.smartevents.manager.v2.services;
 
-import java.util.List;
-
 import com.redhat.service.smartevents.infra.core.models.ListResult;
 import com.redhat.service.smartevents.infra.v2.api.models.dto.BridgeDTO;
-import com.redhat.service.smartevents.infra.v2.api.models.dto.ResourceStatusDTO;
 import com.redhat.service.smartevents.infra.v2.api.models.queries.QueryResourceInfo;
 import com.redhat.service.smartevents.manager.v2.api.user.models.requests.BridgeRequest;
 import com.redhat.service.smartevents.manager.v2.api.user.models.responses.BridgeResponse;
 import com.redhat.service.smartevents.manager.v2.persistence.models.Bridge;
 
-public interface BridgeService {
+public interface BridgeService extends ShardManagedResourceService<Bridge, BridgeDTO> {
 
     Bridge getBridge(String bridgeId, String customerId);
 
@@ -21,12 +18,6 @@ public interface BridgeService {
     ListResult<Bridge> getBridges(String customerId, QueryResourceInfo queryInfo);
 
     void deleteBridge(String id, String customerId);
-
-    List<Bridge> findByShardIdToDeployOrDelete(String shardId);
-
-    Bridge updateBridgeStatus(ResourceStatusDTO statusDTO);
-
-    BridgeDTO toDTO(Bridge processor);
 
     BridgeResponse toResponse(Bridge bridge);
 }
