+++ manager-parent/manager-v2/src/main/java/com/redhat/service/smartevents/manager/v2/services/ProcessorService.java
@@ -1,16 +1,13 @@
 package com.redhat.service.smartevents.manager.v2.services;
 
-import java.util.List;
-
 import com.redhat.service.smartevents.infra.core.models.ListResult;
 import com.redhat.service.smartevents.infra.v2.api.models.dto.ProcessorDTO;
-import com.redhat.service.smartevents.infra.v2.api.models.dto.ResourceStatusDTO;
 import com.redhat.service.smartevents.infra.v2.api.models.queries.QueryResourceInfo;
 import com.redhat.service.smartevents.manager.v2.api.user.models.requests.ProcessorRequest;
 import com.redhat.service.smartevents.manager.v2.api.user.models.responses.ProcessorResponse;
 import com.redhat.service.smartevents.manager.v2.persistence.models.Processor;
 
-public interface ProcessorService {
+public interface ProcessorService extends ShardManagedResourceService<Processor, ProcessorDTO> {
 
     Processor getProcessor(String bridgeId, String processorId, String customerId);
 
@@ -23,12 +20,6 @@ public interface ProcessorService {
     Processor updateProcessor(String bridgeId, String processorId, String customerId, ProcessorRequest processorRequest);
 
     void deleteProcessor(String bridgeId, String processorId, String customerId);
-
-    List<Processor> findByShardIdToDeployOrDelete(String shardId);
-
-    Processor updateProcessorStatus(ResourceStatusDTO statusDTO);
-
-    ProcessorDTO toDTO(Processor processor);
 
     ProcessorResponse toResponse(Processor processor);
 }
