+++ manager-parent/manager-v2/src/test/java/com/redhat/service/smartevents/manager/v2/services/SinkConnectorServiceTest.java
@@ -33,7 +33,7 @@ public class SinkConnectorServiceTest extends AbstractConnectorServiceTest {
     int managedSinkConnectorsTimeoutSeconds;
 
     @Inject
-    SinkConnectorService sinkConnectorService;
+    SinkConnectorServiceImpl sinkConnectorService;
 
     @Inject
     SinkConnectorDAO sinkConnectorDAO;
