+++ manager-parent/manager-v2/src/test/java/com/redhat/service/smartevents/manager/v2/services/SourceConnectorServiceTest.java
@@ -18,7 +18,7 @@ import io.quarkus.test.junit.QuarkusTest;
 public class SourceConnectorServiceTest extends AbstractConnectorServiceTest {
 
     @Inject
-    SourceConnectorService sourceConnectorService;
+    SourceConnectorServiceImpl sourceConnectorService;
 
     @Inject
     SourceConnectorDAO sourceConnectorDAO;
